Increase Connection Limit for Databases again
Quassel for some reason likes to use more connections than it actually needs to the database, having the limit too low leads to Quassel being unable to make any commits or transactions to the tables. In the future I will either remove these lines entirely (if that would be "secure"), or, set them high enough that the limit will never be hit. 5 connections max should be fine for both the User and Database.